네이밍오류수정

+++ src/main/java/com/jyh000223/poligon_backend/dto/DayResultDTO.java
@@ -11,6 +11,6 @@ import java.util.List;
 public class DayResultDTO {
     private int cleared;
     private int total;
-    private List<Integer> clearedGoals = new ArrayList<>();
-    private List<Integer> pendingGoals = new ArrayList<>();
+    private List<Long> clearedGoals = new ArrayList<>();
+    private List<Long> pendingGoals = new ArrayList<>();
 }
